# move checking ptr tracking on item pop into cold helper function