@RalfJung 2022-07-05
@saethlin do you think this makes sense, or should we close the issue without further action?