@RalfJung 2022-07-13
Rebased after @saethlin 's PR landed, and found another cold path in the same `item_popped` function. It now contains barely any code itself. :D (which is probably a good thing)

Before:
```
Benchmark 1: cargo +miri miri run --manifest-path bench-cargo-miri/backtraces/Cargo.toml
  Time (mean ± σ):      9.047 s ±  0.099 s    [User: 8.890 s, System: 0.157 s]
  Range (min … max):    8.957 s …  9.196 s    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path bench-cargo-miri/mse/Cargo.toml
  Time (mean ± σ):      2.470 s ±  0.017 s    [User: 2.233 s, System: 0.239 s]
  Range (min … max):    2.456 s …  2.498 s    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path bench-cargo-miri/serde1/Cargo.toml
  Time (mean ± σ):      3.324 s ±  0.047 s    [User: 3.190 s, System: 0.137 s]
  Range (min … max):    3.262 s …  3.387 s    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path bench-cargo-miri/serde2/Cargo.toml
  Time (mean ± σ):      3.299 s ±  0.019 s    [User: 3.172 s, System: 0.130 s]
  Range (min … max):    3.284 s …  3.331 s    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path bench-cargo-miri/slice-get-unchecked/Cargo.toml
  Time (mean ± σ):      6.613 s ±  0.052 s    [User: 5.982 s, System: 0.633 s]
  Range (min … max):    6.534 s …  6.660 s    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path bench-cargo-miri/unicode/Cargo.toml
  Time (mean ± σ):      4.800 s ±  0.022 s    [User: 4.638 s, System: 0.164 s]
  Range (min … max):    4.768 s …  4.828 s    5 runs
```
After:
```
Benchmark 1: cargo +miri miri run --manifest-path bench-cargo-miri/backtraces/Cargo.toml
  Time (mean ± σ):      8.969 s ±  0.062 s    [User: 8.803 s, System: 0.168 s]
  Range (min … max):    8.876 s …  9.049 s    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path bench-cargo-miri/mse/Cargo.toml
  Time (mean ± σ):      2.461 s ±  0.012 s    [User: 2.234 s, System: 0.229 s]
  Range (min … max):    2.442 s …  2.475 s    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path bench-cargo-miri/serde1/Cargo.toml
  Time (mean ± σ):      3.244 s ±  0.019 s    [User: 3.115 s, System: 0.131 s]
  Range (min … max):    3.217 s …  3.265 s    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path bench-cargo-miri/serde2/Cargo.toml
  Time (mean ± σ):      3.229 s ±  0.044 s    [User: 3.092 s, System: 0.139 s]
  Range (min … max):    3.180 s …  3.281 s    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path bench-cargo-miri/slice-get-unchecked/Cargo.toml
  Time (mean ± σ):      6.248 s ±  0.106 s    [User: 5.618 s, System: 0.628 s]
  Range (min … max):    6.164 s …  6.422 s    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path bench-cargo-miri/unicode/Cargo.toml
  Time (mean ± σ):      4.862 s ±  0.038 s    [User: 4.671 s, System: 0.191 s]
  Range (min … max):    4.819 s …  4.920 s    5 runs
```

This is just on the edge of 1-2σ most of the time, but seems fairly clear on slice-get-unchecked (6% faster). Strangely it seems like a 2σ regression on unicode.

So overall my inclination is to land this? I can't see how it could hurt, anyway...