@saethlin 2022-07-13
Oh yeah this is perfect, `item_popped` inlines now and the hot path through it doesn't call any functions, which is the transformation I was hoping for :+1:

I'm not surprised this is ~perf neutral on most functions and shows the most effect on slice-get-unchecked, because that benchmark stresses this code path more than anything else. A 2σ regression could easily be noise due to the scheduler or some minor perf effect we don't have control over.