@RalfJung 2022-07-14
> item_popped inlines now

Should we add `#[inline]` or even `inline(always)` to to ensure this remains the case, or is that a bad idea?