@saethlin 2022-07-14
I think that `#[inline(always)]` is a bad idea, and `#[inline]` doesn't apply because we aren't trying to enable cross-crate inlining. In my experience applying it when a function isn't being inlined almost always results in slower code, and the way to get the inlining you want is to extract cold paths into `#[inline(never)]` functions as you've done here.

Ideally LLVM could extract cold paths into functions via PGO but I've never seen that happen. And unlike profile data we know what paths are guaranteed to be cold, as opposed to where we just don't have good profiles yet.