@RalfJung 2022-07-14
Okay, fair. I thought `#[inline]` would also shift some inlining weights, not just help for cross-crate things, but I don't really know.

Thanks for the review!
@bors r+