@saethlin 2022-07-14
It does add an inlining hint, but the hint is small, and some people including me and some people in LLVM keep wanting to get rid of it, and the fact that both rustc's implementation of `#[inline]` and C/C++ `inline` do two things is kind of a problem because you can't get the ODR behavior or the cross-crate inlining without the inline hint as well.

And the fact that it adds a hint I think means that it would be worse to add in a situation like this. What I would really like is a `#[rustc::warn_if_not_inlined]` or something like that.